transactions  routes refactored with GPT

+++ sql/schema/004_transactions.sql
@@ -5,8 +5,8 @@ CREATE TABLE transactions(
     stock_symbol TEXT REFERENCES stocks(symbol) ON DELETE CASCADE NOT NULL,
     type TEXT CHECK (type IN ('BUY', 'SELL')) NOT NULL,
     quantity INTEGER NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
-    total_amount DECIMAL(10, 2) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    total_amount DOUBLE PRECISION NOT NULL,
     created_at TIMESTAMP NOT NULL
 );
